+++ src/pages/board/board-page.tsx
@@ -1,3 +1,17 @@
+import { useState } from 'react';
+import Calendar from '@components/calendar/calendar';
+
 export default function BoardPage() {
-  return <>게시판페이지</>;
+  const [selected, setSelected] = useState(new Date());
+
+  return (
+    <Calendar
+      value={selected}
+      onChange={setSelected}
+      markers={[
+        { date: '2025-09-03', tone: 'info' },
+        { date: '2025-09-11', tone: 'info' },
+      ]}
+    />
+  );
 }
